third_party/blink/web_tests/external/wpt/css/cssom/CSSContainerRule.tentative-expected.txt

This is a testharness.js-based test.
FAIL @container name (min-width: 100px) {} assert_equals: containerName expected (string) "name" but got (undefined) undefined
FAIL @container (min-width: 100px) {} assert_equals: containerName expected (string) "" but got (undefined) undefined
Harness: the test ran to completion.
